pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "proofkit"
version = "0.1.0"
description = "Mimik ProofKit - Automated Website Audit QA Tool"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Mimik", email = "hello@mimik.com"}
]
keywords = ["website", "audit", "qa", "seo", "performance", "ux"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "anthropic>=0.18.0",
    "playwright>=1.40.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    # API dependencies
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "black>=24.0.0",
    "httpx>=0.26.0",  # For testing async HTTP
]
openai = [
    "openai>=1.0.0",
]

[project.scripts]
proofkit = "proofkit.cli.main:main"

[project.urls]
Homepage = "https://github.com/mimik/proofkit"
Documentation = "https://github.com/mimik/proofkit#readme"
Repository = "https://github.com/mimik/proofkit"

[tool.setuptools.packages.find]
where = ["."]
include = ["proofkit*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
